+++ packages/net-kit/lib/net_kit.dart
@@ -14,7 +14,9 @@ export 'src/manager/i_net_kit_manager.dart';
 export 'src/manager/net_kit_manager.dart';
 
 /// Params
+export 'src/manager/interceptors/request_extra_keys.dart';
 export 'src/manager/params/net_kit_error_params.dart';
+export 'src/manager/params/net_kit_params.dart';
 
 /// AuthTokenModel
 export 'src/model/auth_token_model.dart';